Using floats again

+++ src/GTAQuat.cpp
@@ -9,11 +9,11 @@ GTAQuat::GTAQuat() : w(1), x(0), y(0), z(0)
 {
 }
 
-GTAQuat::GTAQuat(double w, double x, double y, double z) : w(w), x(x), y(y), z(z)
+GTAQuat::GTAQuat(float w, float x, float y, float z) : w(w), x(x), y(y), z(z)
 {
 }
 
-GTAQuat::GTAQuat(double x, double y, double z) : GTAQuat(vec3(x, y, z))
+GTAQuat::GTAQuat(float x, float y, float z) : GTAQuat(vec3(x, y, z))
 {
 }
 
@@ -31,33 +31,33 @@ GTAQuat::GTAQuat(vec3 degrees)
 
 vec3 GTAQuat::ToEuler() const
 {
-	double epsilon = 0.00000202655792236328125;
-	// double epsilon = glm::epsilon<double>();
-	// double epsilon = 0.0;
-	double temp = 2 * y * z - 2 * x * w;
+	float epsilon = 0.00000202655792236328125f;
+	// float epsilon = glm::epsilon<float>();
+	// float epsilon = 0.0f;
+	float temp = 2 * y * z - 2 * x * w;
 
-	double rx, ry, rz;
+	float rx, ry, rz;
 
-	if (temp >= 1.0 - epsilon)
+	if (temp >= 1.0f - epsilon)
 	{
-		rx = 90.0;
-		ry = -degrees(atan2(clamp(y, -1.0, 1.0), clamp(w, -1.0, 1.0)));
-		rz = -degrees(atan2(clamp(z, -1.0, 1.0), clamp(w, -1.0, 1.0)));
+		rx = 90.0f;
+		ry = -degrees(atan2(clamp(y, -1.0f, 1.0f), clamp(w, -1.0f, 1.0f)));
+		rz = -degrees(atan2(clamp(z, -1.0f, 1.0f), clamp(w, -1.0f, 1.0f)));
 	}
-	else if (-temp >= 1.0 - epsilon)
+	else if (-temp >= 1.0f - epsilon)
 	{
-		rx = -90.0;
-		ry = -degrees(atan2(clamp(y, -1.0, 1.0), clamp(w, -1.0, 1.0)));
-		rz = -degrees(atan2(clamp(z, -1.0, 1.0), clamp(w, -1.0, 1.0)));
+		rx = -90.0f;
+		ry = -degrees(atan2(clamp(y, -1.0f, 1.0f), clamp(w, -1.0f, 1.0f)));
+		rz = -degrees(atan2(clamp(z, -1.0f, 1.0f), clamp(w, -1.0f, 1.0f)));
 	}
 	else
 	{
-		rx = degrees(asin(clamp(temp, -1.0, 1.0)));
-		ry = -degrees(atan2(clamp(x * z + y * w, -1.0, 1.0), clamp(0.5 - x * x - y * y, -1.0, 1.0)));
-		rz = -degrees(atan2(clamp(x * y + z * w, -1.0, 1.0), clamp(0.5 - x * x - z * z, -1.0, 1.0)));
+		rx = degrees(asin(clamp(temp, -1.0f, 1.0f)));
+		ry = -degrees(atan2(clamp(x * z + y * w, -1.0f, 1.0f), clamp(0.5f - x * x - y * y, -1.0f, 1.0f)));
+		rz = -degrees(atan2(clamp(x * y + z * w, -1.0f, 1.0f), clamp(0.5f - x * x - z * z, -1.0f, 1.0f)));
 	}
 
-	// Keep each component in the [0.0, 360.0[ interval
+	// Keep each component in the [0.0f, 360.0f[ interval
 	return vec3(rx, ry, rz);
-	// return mod(vec3(rx, ry, rz), 360.0);
+	// return mod(vec3(rx, ry, rz), 360.0f);
 }
